requirements.txt
fastapi
uvicorn
sqlalchemy 
pymysql
cryptography  
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv